+++ src/black/__init__.py
@@ -1438,6 +1438,22 @@ def patch_click() -> None:
 
 
 def patched_main() -> None:
+    #: Fixes errors with emoji in  Windows terminals when output is redirected
+    # (i.e. pre-commit): https://github.com/psf/black/issues/3156
+    # To be as safe as possible, we only enable this for "Windows 10 October
+    # 2018 Update (version 1809)" and up.
+    if (
+        "pytest" not in sys.modules
+        and platform.system() == "Windows"
+        and tuple(map(int, platform.version().split("."))) >= (10, 0, 1809)
+    ):
+        sys.stdout = io.TextIOWrapper(
+            sys.stdout.buffer, encoding="utf-8"
+        )  # pragma: nocover
+        sys.stderr = io.TextIOWrapper(
+            sys.stderr.buffer, encoding="utf-8"
+        )  # pragma: nocover
+
     # PyInstaller patches multiprocessing to need freeze_support() even in non-Windows
     # environments so just assume we always need to call it if frozen.
     if getattr(sys, "frozen", False):